Move TwitchEmoteOccurrence to TwitchCommon.hpp
